+++ src/core/master-control.ts
@@ -2,9 +2,9 @@ import { startupFuncs, startupCmds, postStartupCommands } from '../core/vim-star
 import { asColor, ID, log, onFnCall, merge, prefixWith } from '../support/utils'
 import { NotifyKind, notify as notifyUI } from '../ui/notifications'
 import CreateTransport from '../messaging/transport'
+import { decode } from '../messaging/msgpack-decode'
 import { Api, Prefixes } from '../neovim/protocol'
 import NeovimUtils from '../support/neovim-utils'
-import { decode } from '../render/msgpack-decode'
 import { Neovim } from '../support/binaries'
 import { ChildProcess } from 'child_process'
 import SetupRPC from '../messaging/rpc'
