+++ src/messaging/msgpack-encode.ts
